requirements.txt
gradio==5.33.0
llama_index==0.12.40
llama_index.llms.mistralai
llama_index.embeddings.mistralai
mistralai==1.8.1
openai
anthropic
PyJWT==2.10.1
python-dotenv==1.1.0
scikit-learn==1.6.1
requests==2.32.3